rtl: add MR descent state before landing
If the corresponding new parameter RTL_DESCEND_MR is positive, the drone
will descend to this altitude after any loitering and transition to MR
(in case of VTOL). This allow for RTL_DESCEND_ALT to be set high enough
for FW loitering, and still descent at a faster rate than landing
speeds in order to conserve battery.

In addition, RTL_DESCEND_ALT and RTL_LAND_DELAY will only be applied for
FW (including VTOL in FW mode), while it will be ignored for MR which
will only use RTL_DESCEND_MR.