drivers: enable IridiumSBD telemetry again
My reasoning is that it's very confusing if everything is just disabled
because there might be more bugs. Everything in PX4 might have bugs with
it but by keepign it disabled we won't surface them so they can get
fixed. However, by disabling it we cause an hour or more of figering out
how to enable it to anyone that does want to try it.

Therefore, my suggestion is to include it again. Of course, it still
needs to be actually enabled via param.

Signed-off-by: Julian Oes <julian@oes.ch>