2ndOrderRef: Enable attitude filtering on manifold
The state can now be an attitude represented by an AxisAngle and the
rate is the body rate.
All the linear algebra and operations on the rates and accelerations are
directly done in R^3 while the attitude additions and subtractions are
performed on the 3-sphere S^3 (i.e.: quaternion group).