mavlink: allow UDP client to reconnect
This handles the case where a UDP client disappears and then comes back
with a different source address.

It's a bit of a weird pattern to connect to PX4 directly to the local
address (e.g. 18570) instead of listening on 14550 or 14540, however, it
can make connections in certain use cases easier, e.g. through docker.

The change is not very pretty and I'm not sure about the side effects
for other UDP setups. It might warrant a broader change.

Signed-off-by: Julian Oes <julian@oes.ch>