Update ecl_yaw_controller.cpp
The ratio of horizontal acceleration to horizontal velocity is the yaw rate. However, the statement "yaw_rate=tan(roll) * cos(pitch) * G/airspeed" in the px4 code calculates the ratio of horizontal acceleration to airspeed, but the direction of airspeed may not be horizontal.

So, this statement should be changed to "tan(roll) * cos(pitch) * G/(airspeed * cos(pitch))", because the denominator "airspeed * cos(pitch)" is the horizontal component of airspeed.

After eliminating the cos(pitch) of the numerator and denominator,  a new formula ”yaw_rate=tan(roll) * G/airspeed“ is obtained.