rc_update: Fix broken negative threshold logic
Previously the negative threshold, which the absolute value of it indicates
the threshold RC channel should be lower than, was being compared
without removing it's sign.

And since the rescaled RC channel value could never be negative, it
could never be lower than the threshold value, leading to the negative
(inversed) threshold function to not work.