[Fixme] Matek H743-slim: Set product ID of USB ACM to be detected by QGC
This sets the product Id to "0x5740", which allows then the board to be
nominally connected to QGC, as the vendor ID and product ID set matches
hte "ArduPilot ChibiOS" item in USBBoardInfo.json of the QGC source
code.

This is a workaround, an a proper product ID and Vendor ID should be
found, but at least users will be able to connect to QGC with this fix.

Note, firmware upgrade with QGC is also broken, as it seems like the
bootloader isn't replying to the "PROTO_GET_SYNC" command, and thus the
Console log shows "getBoardInfo failed, Sync: Send Command: Get Command
Response: Timeout waiting for bytes to be available".

But this can be bypassed by using the px_uploader.py script directly: Tools/px_uploader.py --port /dev/ttyACM0 --baud-flightstack 115200 build/matek_h743-slim_default/matek_h743-slim_default.px4