FixedwingPositionControl: Use constant bank angle on VTOL transition when groundspeed is low and wind invalid
When starting a VTOL transition after takeoff, the groundspeed starts from zero and the wind estimate is potentially not valid yet. In this case the track angle is not yet valid as  well as we don't have a clear direction we are flying to. In this case, the npfg navigation function cannot to much. So use a constant roll angle until the ground speed is increases.