drivers: Accept entire distance sensor signal strength range
Before this commit many valid measurements were discarded as the signal strength was lower than 70/255. Based on our experiments we have only seen outlier measurements when the signal strength is identically equal 0.  Furthermore outliers are removed in the terrain estimator making it not that critical if some outliers are not picked up here.