Added ObstacleDistance publishing from downsampled depth camera pointcloud to test CollisionPrevention in gz-sim.
In this example the FOV of the camera is 73 degrees, the width is 640 and height is 480. I've downsampled this by
8 to get a width of 80 points. I use the middle row from the downsampled point cloud as the input to the ObstacleDistance.distances[]
array. Unfortunately CollisionPrevention in PX4 uses 5 degree sectors, so therefore we can only get 14 points from the 73 degree
FOV (73 / 5 = 14.6). Over the width of 80 points I chose to leave off the first and last 5 points such that we can take the 70 points
and divide them evenly by 14 (70/14 = 5). So our down sampled points end up being from -35 degrees to +35 degrees FOV at 5 degree increments.

To use:
make px4_sitl gz_x500_depth

Co-authored-by: Andrew Wilkins <andrew.wilkins@ascendengineer.com>