logger: log all GPS data
I propose to log GPS data at full rate instead of once a second because
GPS data is usually only 10 Hz anyway, and for debugging every sample
can be an important clue.

For instance, not all messages contain the heading, and so
heading is completely missed and impossible to debug.

Signed-off-by: Julian Oes <julian@oes.ch>