geofence: do not store breach state in geofence class
Since the geofence class is used for detecting a geofence breach, but
also for checking whether a reposition is valid or find a loiter
position after a geofence breach, it can not store breach state. This
would cause errors where reposition outside geofence would not always be
denied, and it could cause the drone to terminate if a reposition
outside a "terminate geofence" was attempted while a less severe
geofence breach was active.