geofence: independent send warning and calulate reposition in navigator
A previous less severe brach action (none or warn) could block
recalcualtion of a new loiter point. Also invalidate the loiter point
when leaving the loiter action geofence, so that the calulated point is
not stored for a later manual loiter if not used (e.g. when an RTL is
active when flying through a loiter geofence).