AirspeedSelector: remvoe @volatile from the ASPD_SCALE params
This flag prevents the params from being loaded from a param file.
Originally it was thought that this scale was more airspeed sensor
specific, but it has shown that it's mainly the position of the
pitot on the vehicle that influences it, and thus it's similar for
a model and should be passed to new vehicles.

Signed-off-by: Silvan Fuhrer <silvan@auterion.com>