Update aux_global_position.cpp
Problem Solved by the Proposed Feature
The existing system only applies the EKF2_GPS_POS_* parameters when vehicle GPS data is used in the EKF. This limitation restricts the flexibility and accuracy of the state estimation process, as other sources of global position data are not considered. The proposed feature solves this problem by allowing the EKF to use auxiliary global position data, enhancing the robustness and accuracy of the state estimation.