ekf2: rename global_origin to ned_global_ref
This variable contains the location of the NED's origin in global
coordinates. It is then not the global origin but the global reference
of the local frame.

Also, to have a valid global position, this reference doesn't only needs
to be initialized but also valid. It is valid when the local frame is
locked in place by using velocity and/or position aiding sensors.