[FREEFLY] rcS: Magnetometer calibration transition mechanisms
rcS: transition factory calibration mag rotation from 23 to 43

Signed-off-by: Claudio Micheli <claudio@auterion.com>

Remove old mag cals from eeprom

Reset mag calibration only if the parameter version is too old

So far, the mag cal is always reset when the current parameter version and the default version from the airframe configuration don't match. This means, the vehicle has to be often calibrated after a software update, even though, the calibration would still be valid.
With the new logic, the calibration is only reset if the parameter version is too old. 61 was the parameter default version in release v1.0.22. The sensor coordinate system in one of the mag drivers has changed in the next release and thus, we still have to reset the incompatible calibration.

Changed threshold for resetting mag cal to all versions before 1.1.11