mavlink_main: stream attitude as quaternion if HIL
Previously with SIH, for some VTOL airframes (e.g. SIH Tailsitter Duo,
1102) the attitude display on the ground station was wrong due to the
body frame in forward flight differing from the usual convention.

This is accounted for when sending the attitude as quaternion [1] over
mavlink, but not when sending it as euler angles [2]. Therefore, in
SIH/HIL we always send it as a quaternion.

1: src/modules/mavlink/streams/ATTITUDE_QUATERNION.hpp
2: src/modules/mavlink/streams/ALTITUDE.hpp