fix: avoid gimbal pan while the vehicle is disarmed
The change in PX4-Autopilot#21010 (commit f887ad6ebf2bda262c1eb9b9c75c891c49c7fc46) allows a yaw
rate with zero throttle. The yaw setpoint is used as a feed-forward for the gimbal and causes the
it to move. Before the change, the gimbal was not moving unless the vehicle was armed and throttle
increased.

With this change, the feed-forward is only set when the vehicle is armed to avoid the gimbal
panning before arming. It still moves on the ground but that is acceptable for now.