Add loiter size and orientation
Params 3 and 4 handling are missing. This adds it in.


Param 3 now will respond to a loiter size and param 4 to 0/1 for CW/CCW. 

Also, on alt only changes if you omit param3/4 it explicitly does what you have set as your default (rather than last done which is how it is now) to ensure you get the default unless the param is passed.