Allocation: expose threshold to decare motor stopped in new param CA_STOP_MOT_THLD
Instead of outputting a command to minimally rotate a motor, it is desirable
to competely stop it if the commanded setpoints is below some value.
E.g. a tiltable motor on a tiltrotor VTOL should be stopped if the commanded
thrust is 0 in fixed-wing mode, but not stopped when in hover (thus PWM_MIN has
to be configured to spin the motor).
This commit replaces the previous hard-coded limit of 2% motor setpoint with
the new parameter CA_STOP_MOT_THLD.

Signed-off-by: Silvan Fuhrer <silvan@auterion.com>