navigator: Cleanup refactor of vehicle_command publishing
Use reference instead of pointer and adhere to minimal conventions for clarity.

These MAVLink dependent interfaces should be pushed out to the MAVLink module
and internal interfaces be independent. A simple request like set gimbal to a neutral position takes three vehicle commands with a specific combination of float and double parameters which have to be filled with gimbal, vehicle and MAVLink context. This should really not be the case.