Gazebo: Added Bypass for MC_RATE_Control=n. Fixes #21248
mc_rate_control=n forces the px4-sitl to build without
mc_defaults, and mc_apps but gazebo-classic still asks for both files
to be generated, leading to error 512.

The body of the message can contain several paragraphs.
Describe in detail what you changed. Link issues and flight
logs either related to this fix or to the testing results
of this commit.

First, added a wrapper file for mc_default. Second,
added if condition so that for both mc_apps and mc_default
files to be generated, if the file is missing, when needed.

Reported-by: karishmapatnaik

Signed-off-by: siri <sriramj2@hotmail.com>