Fix photo triggering when using MAVLink camera on telem1
In some scenarios automated photo triggering during a survey would work (such as when using gazebo-classic_typhoon_h480) but in other siturations it would not (such as when using an external MAVLink camera attached to the telem1 MAV_0 port of a ModalAI voxl). This change makes so automated survey photo triggering works in both scenarios.